Refactor IndividualAction to introduce CollectiveActions
